Integrating to navigation to app for dynamic update of labels
